Improve memo key generation
Summary:
Make some improvements to how we generate memoization keys. The
previous scheme would use various optimized patterns depending on the
parameter's associated type-hint. Not all of these patterns could be implemented
in bytecode, so GetMemoKeyL had different semantics depending on the
type-hint. However, this meant that HHBBC, the JIT, and the interpreter had to
be all kept in sync. Instead, have HHBBC be solely responsible for implementing
specialized memo key patterns. HHBBC will lower the GetMemoKeyL opcode to the
specified generation scheme if possible. Therefore, if GetMemoKeyL remains, the
JIT and the interpreter can simply treat it using the generic scheme (rather
than peeking at the type-hints and interpreting it differently).

To enable this lowering in all the necessary ways, introduce two new
bytecodes. The first, Select, represents a simple ternary operator, and is used
for lowering where we want to decide between two values based on some
predicate. The second, DblAsBits, is used to reinterpret a double value as an
integer (with the same bit-pattern).

As a side-effect of this, the bug where a deprecated memoized function would
crash in HHBBC (when DisallowDynamicVarEnvCalls is disabled) is now fixed.

Finally, even for the generic GetMemoKeyL opcode, if the JIT knows the input
type, call the appropriate serialization function directly, rather than calling
a generic helper which checks the type and then dispatches.

Reviewed By: markw65

Differential Revision: D8302966

fbshipit-source-id: c45dec0fdf3f0b39b532f906f786bf308a582b93