Add F14 memo cache implementation
Summary:
Add the new implementation of memo caches using F14. This just adds the
implementation and interfaces but doesn't actually use it in the runtime. That
will come in later diffs. This uses templates to specialize the memo caches
according to key count and key types, which for the common cases, generates very
compact and efficient code.

Reviewed By: markw65

Differential Revision: D8307445

fbshipit-source-id: 136c98c0e31a7f9dde4056b57164e2460c6abfae