Add runtime support for new memoization storage
Summary:
Add support in the runtime for storing the new memoized data. Right now, memo
data is stored in either static locals, static properties, or instance
properties, just like any other PHP value. With the new scheme, memo data for
static functions is stored in RDS. For instance methods, memo data is stored
*before* the object. The memo data is either a Cell (for functions with no
parameters), or a pointer to MemoCacheBase (for functions with parameters).

As stated, memo data for instance methods is stored before the object. This is
to avoid colliding with properties and ensures that once a class allocates a
slot for memo data, that slot's offset never changes. Similar to NativeData,
invent a new header kind to represent the start of the memo data. For objects
with native data, store the memo props before the native data.

The actual memo slots are assigned when the Class gets loaded. Since its
beneficial to use non-shared caches (especially when the function takes no
parameters), we use a heuristic, where we'll assign up to 10 (which is
configurable) non-shared slots, then use shared slots for the rest. We give
preference to functions which take no parameters. We need some limit because
Class with a lot of memoized functions will get allocated too many slots.

This diff only adds the support for assigning and allocating the new memo
slots. Nothing in the runtime actually uses them (yet), and the old memo scheme
is still what's used at runtime.

Reviewed By: markw65

Differential Revision: D8325385

fbshipit-source-id: e919262b584c8e3cb7bf14456b9c5f737ba5f06b