Change semantics of LocalRange immediate
Summary:
When LocalRange was originally defined, it was considered undesirable to allow
for it to represent the empty range. This is because it was used in opcodes that
made no sense with empty ranges and it seemed better to forbid that case by
construction. So, the second field of the range was a count of the locals
besides the first one (so a count of 0 meant there was only one local).

However, the new memo bytecodes will work with empty ranges, so its time to lift
this restriction. Change the count field of LocalRange to have the more
intuitive meaning of all the locals (so 0 means an empty range). If the count is
0, the first local is arbitrary. This mainly involves just removing - 1 from a
few places and changing < to <=.

Reviewed By: jano

Differential Revision: D8328347

fbshipit-source-id: f89c3d75fc18723adb979efb2faab14ae408ad10