Switch over to new memoization machinery
Summary:
Redefine MemoGet and MemoSet to use the new memoization machinery and
remove the now unneeded old machinery. The new method is considerably simpler
from the perspective of the emitter. MemoGet and MemoSet will now automatically
take advantage of the type of function, as well as the number of parameters, so
the emitter does not need to handle so many different cases. No special
properties need to be created (so no more guarded versus unguarded caches, etc).
The special sequences involving IsMemoType and MaybeMemoType are gone as
well. IsMemoType, MaybeMemoType, and IsUninit are now unneeded and are removed.

MemoGet now takes an offset immediate and a local range immediate. Its
responsible for doing everything necessary to lookup the appropriate memoization
value. If its not found, it branches to the specified offset. This removes the
need for a special sentinel value (like uninit) to mean a value wasn't found
(and a redundant check). This bytecode is somewhat novel in that it can either
branch or push a value, which required a little bit of adjustment in places
where we track the stack (like the emitter, assembler, and HHBBC emitter).

Also implement clear_static_memoization and clear_instance_memoization can be
used to clear memoization data. (even in repo mode!) The old style way of using
reflection will no longer work, as the memo data is no longer stored in PHP
visible properties.

While I was at it, I noticed that HackC wasn't dealing with memoized functions
being marked deprecated properly, so fix that.

Reviewed By: markw65

Differential Revision: D8337098

fbshipit-source-id: 17e79c0cf806d007373baf8bb7c009894a160c7f