Fix SIMD reg to reg copy on ARM
Summary:
This diff changes copy to use a Neon mov instruction (ORR).  Since all of the
vector support in our forked vixl assembler has been jerry-rigged in (D3221842),
I simply added the support for the single instruction needed.  In the upstream
version of vixl, they have full support for the neon instructions at this point.

It remains to implement disassembly support for this neon instruction, and the
others added in D3221842.

Reviewed By: mxw

Differential Revision: D8475230

fbshipit-source-id: 1ddd65e8846abb5e412dbde616426fc58aa6ebcb