Delete pcntl_wait test
Summary:
remove test that relies on particular set of child processes.
`pcntl_wait` waits for any child of current process. Prior to HackC the only child process was spawned in test and at the point of calling `pcntl_wait(WNOHANG)` it already exited so this call returned -1. With HackC there is a pool of extern compiler processes that is created before running tests so even after child process created in test finishes there are still child processes (compilers in pool) that are alive and `pcntl_wait` returns 0 indicating that.

Reviewed By: alexeyt

Differential Revision: D8500994

fbshipit-source-id: 0022f5b59a84af74dcc72b4f64be3315315e8fe6