Consider ?bool to be sketchy in null checks
Summary: Currently, we don't warn when ?bool is used in a sketchy null check. One might argue that the behavior is clear enough that it's reasonable for us to be permissive (i.e., both false and null are considered falsy), but it's probably better to reject this for the sake of consistency.

Reviewed By: DavidSnider, andrewjkennedy

Differential Revision: D8407897

fbshipit-source-id: d32219f72a03767982f8aa900b13d2e016521bcc