Don't allow function use clauses to shadow global names
Summary: The no_global_fallback2 test verifies that we emit an error when a function use clause attempts to shadow a function in the global namespace. Previously, we were only emitting redefinition errors for functions defined in the same namespace. After this change, we keep track of the set of defined global function names across NamespaceBodies, so we are able to emit an error when a function use clause attempts to shadow one of them.

Reviewed By: oulgen

Differential Revision: D8372239

fbshipit-source-id: 77cf80ef9833192439bcbff71a2b64a72b84416f