Revert nonsensical change from D6911678
Summary:
In D6911678 for some reason hackc started trying to parse the body of a vect, dict, or keyset literal as an expression, and if it _didn't_ parse as an expression we would lower the code as a vec/dict/keyset literal even if we were in a PHP file and enable_hh_syntax was off. This makes no sense.
Additionally, the code used exceptions as flow control to accomplish this, which is fragile.
Revert the problematic part of D6911678.

Reviewed By: oulgen

Differential Revision: D8506177

fbshipit-source-id: 0807abe0c924a5b8735bc9129a45cc504fa42107