class_meth with generic class
Summary:
`class_meth(C::class, 'method_name')` doesn't work properly if `C` is a generic class.

Fix is to instantiate generic parameters of `C` with type variables that will unify on use; also, make sure that constraints on the parameters are satisfied.

(Note that a more expressive design would generate a *generic* function type that could be used on different instantiations of the generic parameters. But we don't support this notion.)

Reviewed By: manzyuk

Differential Revision: D8504223

fbshipit-source-id: ead9725e1eca518918bbe2f94eeb3c931cf94031