Raise parsing errors more like php -l does
Summary:
Before this diff, if we found an error in the parse tree, we would arbitrarily take the last one and show it to the user.

We should always run Full_fidelity_parser_errors if we've seen any errors in the parse tree.

This brings us closer in line with how php -l raises errors, which should improve our error messaging greatly.

Reviewed By: oulgen

Differential Revision: D8543409

fbshipit-source-id: 393d436634299b2c27bb0ad8e429fe6938034225