Use FFP for typechecker tests, part 1
Summary:
This diff moves all toplevel tests that have a difference from the FFP to the legacy parser into their own folder. There are 400 of them, the majority of which are positional differences which I'll clean up in a different diff.

There are also 300 or so discrepencies among files in subfolders, but I'll tackle the subfolders one at a time.

Reviewed By: kmeht

Differential Revision: D8556948

fbshipit-source-id: 7bb73163db7eba2727c93bb36bdf0f616ee9d666