Add is_any_array builtin for idx
Summary:
This goes a long way to reducing the noise from enabling is_array
notices.

Since the JIT will pessimize `idx` to calls to `HH\idx` if it can't
deduce the first argument's type well enough, we need to make sure that the
systemlib implementaiton of `HH\idx` doesn't call `is_array`. The check
previously here was equivalent to the current www `is_any_array` and we don't
currently have a way (in user space) to check for array-ness inclusive of
varrays, darrays, vanilla arrays, and all hack arrays.

Reviewed By: ricklavoie

Differential Revision: D8527970

fbshipit-source-id: 1bff8ff494af580ae0df2f8ed861a6ea2bc54a6f