Delete unused IncStat bytecode.
Summary:
Nothing emits the IncStat bytecode, so delete it.
RuntimeOption::EnableEmitterStats used to control whether to emit it, it's dead too, so delete it also.
The IncStat IROP is not dead, however, so keep it.
The IROP is being used via the emitIncStat helper in irgen-basic.cpp, but that is not being declared in irgen.h any more since the bytecode is gone. Add an irgen-basic.h to declare it from and do so.

Reviewed By: swtaarrs, jano

Differential Revision: D8587685

fbshipit-source-id: eadd32c87945ec480c3f7e04c86a654163c9c400