Fix MSRV inout return sequences in HackC
Summary: The implementation of RetM in HackC was buggy and would emit VerifyOutParam instructions after IterFree when returning from inside a loop, which could cause us to double free an iterator. We would also emit the IterFree following the RetM in some cases, making it effectively useless and allowing us to leak iterators.

Reviewed By: jano

Differential Revision: D8601061

fbshipit-source-id: 1fae221f513950a6152287c884c43f42596c050d