Emit using blocks even if position information is missing.
Summary:
When we emit using blocks, we (as of D7137237) try to compute the range of positions that are covered by the using block. Unfortunately, we have a number of syntax elements where we punted on providing position information (Pos.none), and if one of those (e.g. an UNSAFE BLOCK) is at the start or end of the using block then the Pos.btw operation calculating the range will fail.
Handle this gracefully by degrading to using just the start or just the end position if only one is available, and to Pos.none if neither is available.

Reviewed By: oulgen

Differential Revision: D8601053

fbshipit-source-id: a4e127deb9dc859ea6546c65952279c6e3939927