Format position of character according to LSP
Summary:
Changes have been made earlier to ensure that Nuclide was sending the correct position according to LSP - this means one position immediately to the right of the character typed.

The Hack server was modified to account for versions of Nuclide that were sending the correct position and versions that weren't, and to send the correct position regardless.

Recently, we found that the check we placed in the Hack server still ended up giving the wrong position for certain formatting cases (T30616028).

However, this also means everyone's version of Nuclide should be up to date now and that as described in D8227541, it is now time to remove the check so that we only return the previous position (AKA the position of the character itself).

Reviewed By: arxanas

Differential Revision: D8580705

fbshipit-source-id: f98ef37561d3fc44294049d0d7cfc680e8a9baab