Conditional reactivity
Summary:
Prior to this diff functions can be conditionally reactive only for the parameter of the receiver. 'Conditionally reactive' function with condition type `T `here means that function can be invoked in reactive context if static type of the receiver is known to be a subtype of `T` or have a condition type `T2` which is a subtype of `T`. With this diff condition for conditionally reactive function can be attached to any parameter of the function - for example
```
<<__Rx, __OnlyRxIfArgs
function f(<<__OnlyRxIfImpl(Rx::class)>> A $a): void { ... }
```
here call to function `f` is allowed in reactive context only if argument that correspond to parameter `$a` is s subtype of `Rx` - receiver is treated similarly to any other parameters.

In order to  associate condition type with parameter type (`Rx` with `A` in example above) we do the following: when localizing function parameters if parameter has associated conditional type we'll create a generic type with name uniquely derived from the pair (declared type + condition type). This name is used as a key for the map in `env` that maps name of generic parameter to a condition type. Then we set declared type as a constraint for this fresh generic type and return generic type as type of parameter. This allows to retrieve condition type for a given parameter type later when we do checks for function calls.

Reviewed By: jamesjwu

Differential Revision: D8601706

fbshipit-source-id: 4169dace804d63a8596fefab5e0166e7799bd1d4