Don't save HH_FIXME in quick mode.
Summary: When running in quick mode, we drop function bodies. This means that when we go back and try to find out all of the HH_FIXMEs in the file that we just parsed, there are none!  Or rather, there are very few.  Instead, we should not save HH_FIXMEs if the file was parsed in quick mode, so that we know to go back and read them later.

Reviewed By: dabek

Differential Revision: D8630554

fbshipit-source-id: 06d4e3fed6664a7529637a9c293a82e063ca9c9b