Improve typing of arrays containing untyped values
Summary: When typing an array literal, if one of the values is untyped (i.e., has type the "any" type `_`), then we consider the entire array untyped (i.e., give it the value type `_`), but we do not check if the typed elements of the array have the types compatible with the expected type passed down from the context.  This diff fixes this.

Reviewed By: CatherineGasnier

Differential Revision: D8537974

fbshipit-source-id: e9ff6df8613bddcd10b965448508be5226d2a49b