Separate decl vs. regular HH_FIXMEs
Summary:
This creates two heaps insetad of one for HH_FIXMEs: DECL_HH_FIXMEs and HH_FIXMEs.

We put HH_FIXMES into DECL_HH_FIXMEs if we quick parse, and regular HH_FIXMEs if we full parse.

When we look for the presence of an HH_FIXME, we check the HH_FIXME heap first, followed by the DECL_HH_FIXME heap.

We clear both heaps entries whenever a file changes.

Reviewed By: pittsw, arxanas, dabek

Differential Revision: D8666452

fbshipit-source-id: 8d43b04c74259992f0cecf1b2ff084682609948d