Fix return values for PDOMySqlConnection::commit/rollback
Summary:
Between mysql-5.x and mysql-8.x the return values of mysql_commit and
mysql_rollback changed from my_bool (an int) to bool.

According to the docs, in both cases non-zero means failure, and zero
means success (its not clear whether the old behavior of -ve for
failure was correct in practice, even for 5.x).

I found these because gcc-7 reports bool >= 0 as a tautology.

Reviewed By: alexeyt

Differential Revision: D8642138

fbshipit-source-id: af638bc9612b36a963afca2cf3cfe41f51201f66