Fix VerifyRetNonNullC
Summary:
We would kill it if !is_opt(stackT); but eg neither TInitCell nor
TInitNull is_opt, so we would kill it when it might fail, and we could
kill it when it would definitely fail. We can also set unreachable
when its known to fail.

In practice, this probably wasn't breaking anything, because we only
create VerifyRetNonNullC when we have an is_opt type, and types should
only get more refined; so the only possible failure mode would be an
OptFoo becoming an InitNull later, and getting optimized away (and
thats pretty unlikely).

Reviewed By: mofarrell

Differential Revision: D8678459

fbshipit-source-id: d205e2e376501c2b4827b7cad8fee260e6ff5f25