Move AttrNumArgs from Attr to a bit in Func::SharedData
Summary:
AttrNumArgs is not being used by the emitter or by hhbbc, so we can move it into Func::SharedData to make room for the emitter to set more bits.
Do so by parsing it to a Native::Attr instead of an Attr in FuncEmitter::parseNativeAttributes, and having FuncEmitter::create stash it in the SharedData of the func it's creating.

Reviewed By: markw65

Differential Revision: D8676334

fbshipit-source-id: 95ff87446ddeba61a1573148892b52f1dbb12b5a