Error for ambiguous object access
Summary:
Consider a class with a private field.

Within a method of this class, the field should be accessible for all objects of this class, which includes its subclasses.

When carrying a type hint of its subclass, the typechecker incorrectly believes that the private field cannot be accessed. This does not reflect the behaviour of HHVM and potentially creates a type hole where the subclass has a public field with a different type.

The same rule applies for private methods.

The fix is to prioritise the private fields when inside a method and the object being accessed can be upcast to the current class, so that the type inference is reflecting the runtime behaviour, but emits an error so that the user is aware.

i.e.
```
class A {
  public function __constructor(private int $i) { }
  public function foo(B $b): int {
    return $b->$i; // Gets the private int in A instead of public string in B
  }
}

class B extends A {
  public function __constructor(public string $i) {
    // some init
  }
}
```

After digging some code in HHVM implementation, we can confirm this behaviour at
diffusion/FBS/browse/master/fbcode/hphp/runtime/vm/class.cpp;84d62943594088bc5c80b93f571b493f3120d8ab$1003-1015

Reviewed By: arun11111

Differential Revision: D8121626

fbshipit-source-id: 0ed70d62035fa43eafe4c4e759196c79d208de76