FFP support for prefixed strings
Summary:
Punting on qualified names and whitespace for now.
______________

Added a new full fidelity syntax kind called PrefixedString with a name field (for the prefix) and a str field (for the string that is being prefixed).

When a double-quoted string occurs after a name, the FFP will now parse it as a PrefixedString.

The RFC is mainly re: whitespace. Ideally, we'd have `re"blah blah` and not `re "blah blah"` parsed as a prefixed string. This version of the diff recognizes both as prefixed strings and fails a test I wrote (test_prefixed_string_errors_2). I'm having trouble finding a way to recognize whether there is whitespace between the Name token and the next token, which must match with a DoubleQuotedStringLiteral or a DoubleQuotedStringLiteralHead token before the whole thing (name and string) gets parsed as a prefixed string (see full_fidelity_expression_parser.ml).

Reviewed By: kmeht

Differential Revision: D8554260

fbshipit-source-id: cf99447f7e981e0d7ec0080852d931c60ff61778