Convert dynPropTable to req::fast_map
Summary:
dynPropTable needed reference stability because a few things
held pointers to the ArrayNoDtor value in the map across operations
that mutated the dynPropTable itself. It seems reasonable to hold
references to ArrayNoDtor for short periods, because Array has all
the logic for copy-on-write and resize escalations of the underlying
ArrayData.

But the map can be significantly more efficient by not allocating
a separate object for every pair of ObjectData+ArrayNoPtr, so fix
the few places that held references across map mutations, and tweak
a few apis to make it safer to use, and convert it to req::fast_map.

Reviewed By: markw65

Differential Revision: D8395329

fbshipit-source-id: 39cd4ac1b337721f7464df8e6727177c59f9f1d7