Add missing hphp_opts file for builtin_coercion_warning.php
Summary:
These warnings only happen when we don't use FCallBuiltin, so we need
to prevent that in repo mode.

Currently Eval.EnableCallBuiltin is broken in our tests, because its
not passed through the global data - so it only works when hhbbc is
run as part of hphpc. Use JitEnableRenameFunction instead.

Reviewed By: ricklavoie

Differential Revision: D8721675

fbshipit-source-id: ef8719ea36918e755924a5f29b610fb2ad58d2f9