<<__PPL>> Obj call typechecking
Summary:
Methods on <<__PPL>> classes are rewritten to be coroutines (with the exception of the constructor). Furthermore, because there is no type information during codegen phase, the ability to perform type based rewrites is limited. As such, it is only possible to correctly rewrite PPL calls that have the `$this->...` format. As a result, we should ban all of the other incorrect ways to call <<__PPL>> methods (that go through the objcall path).

This includes: Calling methods on PPL classes from non ppl calls and functions. Calling methods on PPL classes not using $this->.... and Calling methods on PPL classes using $this->... but in the constructor of the PPL class.

More specifically:

Added an environmental flag for inside ppl class.
Added an error.
Added a check through the obj call typechecking path.

Reviewed By: michaeltingley

Differential Revision: D8099562

fbshipit-source-id: 19d916dc93b855b8aee94b33f283e65fbdae0c20