<<__PPL>> static method call typechecking
Summary:
Methods on <<__PPL>> classes are rewritten to be coroutines (with the exception of the constructor). Because there is no type information during codegen phase, the ability to perform type based rewrites is limited. It is only possible to correctly rewrite static PPL calls based on the self::, parent:: or static:: formation. As a result, we ban all other ways of calling static methods on PPL classes.

Typechecker will now error when:
Static PPL class methods are called without using self::, parent::, or static::
Static PPL class methods are called from within the PPL class constructor

Reviewed By: michaeltingley

Differential Revision: D8103892

fbshipit-source-id: dc4bc9b8921895bc7d1ce1fe6220aa2b25d4ecfc