<<__PPL>> Special method typechecking
Summary:
There are four functions that are specially handled within PPL class methods. These are `sample`, `factor`, `observe` and `condition`. The goal is to treat these four functions as if they were methods on the `Infer` class.

At the time of naming, when we spot a function named one of these, because we do not have context, we will simply bypass naming checks until typing phase. Unbound errors will ultimately still be caught but they will not be naming errors, they will be type errors.

Reviewed By: michaeltingley

Differential Revision: D8107331

fbshipit-source-id: 440fa612074bfcf38652191d8179a417506819ea