Use the same extraction scheme as HackC for embedded data
Summary:
Change the embedded data extraction scheme for type-scanners and
member-reflection to match HackC. Rather than always writing to a temporary file
in /tmp, use three new runtime options: EmbeddedDataExtractPath,
EmbeddedDataFallbackPath, and EmbeddedDataExtractTrust. If
EmbeddedDataExtractTrust is set, than any existing file at
EmbeddedDataExtractPath will be used. Otherwise, the contents of the embedded
data are extracted to EmbeddedDataExtractPath, or EmbeddedDataFallbackPath if
that fails.

Introduce a new piece of generated build-info called a "build-id". This is a
hash of the hhvm_link executable, along with any shared-objects (like the
type-scanners) which will be embedded into the executable. Unlike other
generated build-info, it cannot be overridden. Since its generated from the
executable's binary contents, it uniquely identifies a particular build
(including compile options), not just a source control revision.

EmbeddedDataExtractPath includes the new build-id so is automatically tied to
the HHVM build. EmbeddedDataFallbackPath does as well, except it includes
placeholders for a randomly generated name, and is automatically cleaned up when
the process exits.

Since we'll use EmbeddedDataExtractPath most of the time, and that file will
persistent across runs, this should greatly reduce the chance of leaving files
in /tmp on unclean shutdowns.

Reviewed By: markw65

Differential Revision: D8710487

fbshipit-source-id: c5b68796cdd292abf2a432371a9b7a4609a4f310