Elide Tany when flattening unresolved types
Summary: We can end up with `Tany` as part of an unresolved type e.g. when pushing the option out of an unresolved type (e.g., if one of the unresolved types is `?^_`, which is the type of `null`).  It can be omitted and the resulting type will unify with the original type.  However, fltering out `Tany` allows us to detect more errors because we have check that rely on a particular form of an unresolved type.

Reviewed By: kmeht

Differential Revision: D8695266

fbshipit-source-id: 2083e2e5809a2e23f1e327b355126181bc2e9e71