Run closure convert over all expressions on the AST
Summary:
This was all broken due to a catch all match wildcard...

Before this diff,

```
$x = 1;
() ==> { return $x is int; }();
```
would result in undefined variable as we didn't capture $x from the outer scope. This diff fixes it for is/as expressions as well as all the remaining nodes on the AST that didn't have a closure convert step.

Reviewed By: kmeht

Differential Revision: D8723198

fbshipit-source-id: 8a0021f797eb1d9baf8659d92ff23e4964ebf1bc