Temporarily ban is/as-expressions in lambdas
Summary:
Due to a HackC bug in which we weren't always running closure convert over all expressions in the AST, is/as cannot be used with variables that are captured in a closure.

In the interest of moving fast, this blanket bans all is/as-expressions in lambdas, with an error message indicating that users can FIXME it if they're not using a captured variable.

Reviewed By: vassilmladenov

Differential Revision: D8725780

fbshipit-source-id: 1e9c6a24f5d7c110fcce9342a98096470cc5359c