Improved assembler exception handling
Summary:
Let FatalErrorExceptions propagate out of the assembler/compiler - we want e.g. OOMs and timeouts to kill the request thread without doing any more work.
Make any other exception leaving the unserializer into the assembler a subclass of AssemblerError, so that we can treat them as an internal compiler error and not create a fatal unit.
The expectregex for hphp/test/slow/parser/memory_exhaust.php had to be changed in 2 ways: the exception type in the hackc case changes, and in hackc repo mode it now produces a totally different error, since the initial compilation did not produce a unit.

Reviewed By: markw65

Differential Revision: D8728711

fbshipit-source-id: 4995fca4cd1c194ba9ce0da8b1d4e0d6365059f3