Copy repo-auth-types for public static properties into derived classes
Summary:
HHBBC can infer types for public static properties, but we weren't
copying the repo-auth-type into derived classes when constructing properties, so
we couldn't take advantage of it. So, if class A defines a public static
property called "x" and class B extends A, then A::$x would take advantage of
the repo-auth-type, but B::$x wouldn't (even though its the same static
property).

Reviewed By: markw65

Differential Revision: D8731038

fbshipit-source-id: 07ee16f1c3cbacaafa8682db3eb28d22806c607c