Add a newline after the <?hh declaration during rewriting
Summary:
Currently code such as
```
<?hh
// Copyright ...
```

is rewritten to

```
<?hh // Copyright ...
```
during coroutine lowering.

This is problematic as the copyright comment in the lowered code is parsed by the FFP as the Hack mode. D8691127 will make this a parse error as only valid Hack modes will be allowed in the comment.

To fix this, I've updated coroutine lowering to add a newline after the `<?hh` declaration for non-decl files (the `<?hh` declaration is not rewritten for `decl` files).

Differential Revision: D8713984

fbshipit-source-id: 6b7f7ab6b373c4db1b1c5f03b0efe0e94211f176