avoid fatal with illegal '<' in xhp body
Summary:
When parsing an xhp body, the `scan_xhp_body` will scan until it sees a '<' character, but then the code will `assert_token LessThan`, when the next token might be a `<=`, `<<`, `<=>` and all that.
So we use a cleverer verstion of `assert_token` which accept a tokenizer function, and use `next_xhp_body_token` instead of `next_token` as the tokenizer.

Reviewed By: kmeht

Differential Revision: D8716143

fbshipit-source-id: adb7fd70b4ee7b83460f4bb48ffb2eecf195f2a2