Don't validate constraints of wildcard type parameter
Summary:
Type aliases are tricky to validate because an invalid type could be hiding anywhere in the localization process. That's why we validate every intermediate decl type during localization with the `validate_dty` function on the `ety_env`.

However, once we get to a wildcard type parameter, we no longer need to verify the type, since we'll never validate the type parameters at runtime. Note that we will still raise an error if a user tries to use the `is` operator directly with a type alias with a generic (see D8148890), or if they try to bind the type parameter themselves.

This simply removes the validator when localizing a wildcard type parameter's constraints.

Reviewed By: andrewjkennedy

Differential Revision: D8733406

fbshipit-source-id: 63258ee7386e56265e0256f763ce832e82d74bee