Use namespace aliases in hhconfig in autocomplete
Summary:
In addition to hardcoding the HSL, we should also use the namespaces aliases defined in the hhconfig to populate the list of namespaces in the codebase.

This allows us to thoroughly improve the namespace-aliasing autocomplete test, which will allow us to solve the autocomplete problem in T22617428 with more confidence.

Reviewed By: pittsw

Differential Revision: D8519956

fbshipit-source-id: abe7977cc309e1f2335d10f64f33bf703d006ae4