We should supprss warning 52 by [@@@warning] not by TARGETS
Summary:
Warning 52 comes when you pattern-match directly on the constructor argument of `Sys_error`, or other types which are declared with the annotation "please don't pattern-match on my argument since it's for informational purposes and subject to change".

Well, (1) there aren't any better alternatives, (2) all I want to do is move suppressions out of TARGETS files and into the .ml files. That's what this diff does.

Reviewed By: arxanas

Differential Revision: D8720624

fbshipit-source-id: c53a92c405c52acaf2b03569300d2b5da7b689ae