Disallow arbitrary expressions on the right hand side of instanceof.
Summary: PHP seems like it would allow arbitrary expressions on the right-hand side of an `instanceof` operator, but it actually only allows a subset of expressions, referred to as a [class-type-designator](https://github.com/php/php-langspec/blob/master/spec/10-expressions.md#instanceof-operator).  The FFP allows arbitrary expressions, but hphpc will throw a fatal parse error if the expression on the right isn't a subset of `class-type-designator`.  In order to avoid people checking in code that will crash HHVM, we need to bring the FFP in line with what hphpc will parse.

Reviewed By: oulgen

Differential Revision: D8429596

fbshipit-source-id: a04ae463f95a51790646eb163448553e528a73c0