add --ignore-type-errors flag
Summary:
add ignore type errors flag for knowingly producing
saved states with type errors in them.

This simplifies local testing so it doesn't require a rebuild.

Eventually we're going to start producing saved states with potential type errors
in them and having this as a CLI will help with that.

Reviewed By: ljw1004

Differential Revision: D8741315

fbshipit-source-id: 95946294c87ef655d2f09d236fcd8b1e1c275855