Add subtypeOf(trep) and couldBe(trep)
Summary:
subtypeOf(T<x>) (and couldBe) can be determined solely by looking at
the trep of T<x>, but in practice the compiler generates code to load
the trep, and then check the tag to make sure it doesn't have data
associated with it. Although we've streamlined this, its still much
better to just pass the trep in directly.

This also means that we can often do `subtypeOf(BX | BY)` rather than
`subtypeOf(BX) || subtypeOf(BY)` (even when BX | BY isn't a
"predefined" type).

In addition, add a `subtypeOrNull(bits)` method which is equivalent to
`subtypeOf(bits | BInitNull)`; this will be changed to BNull in later
diffs when we add support for BUninit in optional types (and avoids
the need to add named types for all the new nullish types that will
arise).

Reviewed By: ricklavoie

Differential Revision: D8714389

fbshipit-source-id: 83f548f8afbcbeddb4f51384e0d0c9948ea3082c