Add error for invalid Hack mode comment
Summary:
The legacy parser would raise an error if an invalid mode is specified for a Hack file, e.g. the following would result in a parse error:

```
<?hh // wrong-mode
```

The FFP doesn't raise errors in these cases.

This diff adds an error pass to the FFP that raises an error if an invalid mode is given for Hack files.

Reviewed By: CatherineGasnier

Differential Revision: D8691127

fbshipit-source-id: cb41382400e43d7dcf543a720d79fd89745092e8