bug in `current_token_text`
Summary:
`current_token_text` used to peek a token, get its width, then get the token text by taking the substring of the source text from the start of the lexer with the width of the token. However, the start of the lexer at this point is the start of the leading trivia, so usually the space before the token, therefore we would get strange error messages like the ones which are fixed in this diff.
Instead, here we use simply `Token.text`, which require to add a dummy `text` function to `Full_fidelity_minimal_token`. This is not completely satisfactory because dummy functions are bug-prone, but there is already a dummy `leading_start_offset` in `Full_fidelity_minimal_token`, and doing it the old way, trying to guess where is the start of the token, would be tricky.

Reviewed By: kmeht

Differential Revision: D8732601

fbshipit-source-id: 1473c16e6aac603bb90db5f4d899378ee2b13c70