Refine the type of as-expressions
Summary:
Previously, we'd refine the type of only variables that were used in as-expressions. However, we should also be refining the type of the expression itself, because it may be used where the refined type is required.

The tests have a lot of temporary errors for dict/vec, which will be regenerated once the HHVM fixes are released. The keyset test is representative of the final state.

Reviewed By: CatherineGasnier

Differential Revision: D8745376

fbshipit-source-id: 2735e6fa322c34d19ef9d08eb80ffb5f26ec7f20